requirements.txt
numpy
opencv-python
scikit-image
tqdm
matplotlib
tensorboardX
pytorch_lightning
timm
einops
pillow 